Rename result classes to follow more generic OrderResult / OrderError convention and move them into core.
