--- Card/src/main/java/com/paypal/android/card/ConfirmPaymentSourceError.kt
@@ -1,7 +1,0 @@
-package com.paypal.android.card
-
-data class ConfirmPaymentSourceError(
-    val name: String? = null,
-    val details: List<ConfirmPaymentSourceErrorDetail>? = null,
-    val message: String? = null
-)
