+++ Card/src/main/java/com/paypal/android/card/ConfirmPaymentSourceResult.kt
@@ -1,6 +1,9 @@
 package com.paypal.android.card
 
+import com.paypal.android.core.OrderData
+import com.paypal.android.core.OrderError
+
 data class ConfirmPaymentSourceResult(
-    val response: ConfirmedPaymentSource? = null,
-    val error: ConfirmPaymentSourceError? = null
+    val response: OrderData? = null,
+    val error: OrderError? = null
 )
