--- Card/src/main/java/com/paypal/android/card/ConfirmedPaymentSource.kt
@@ -1,9 +1,0 @@
-package com.paypal.android.card
-
-data class ConfirmedPaymentSource(
-    val orderID: String?,
-    val status: String?,
-    val lastDigits: String?,
-    val brand: String?,
-    val type: String?
-)
