+++ Core/src/main/java/com/paypal/android/core/OrderData.kt
@@ -1,0 +1,6 @@
+package com.paypal.android.core
+
+data class OrderData(
+    val orderID: String,
+    val status: OrderStatus
+)
