+++ Core/src/main/java/com/paypal/android/core/OrderError.kt
@@ -1,0 +1,7 @@
+package com.paypal.android.core
+
+data class OrderError(
+    val name: String? = null,
+    val details: List<OrderErrorDetail>? = null,
+    val message: String? = null
+)
